feat: dbへの登録を実装
