+++ db/0001_plain_morbius.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `repositories` DROP COLUMN `libraries`;
