+++ db/meta/_journal.json
@@ -8,6 +8,13 @@
       "when": 1756187699122,
       "tag": "0000_init",
       "breakpoints": true
+    },
+    {
+      "idx": 1,
+      "version": "6",
+      "when": 1756355816761,
+      "tag": "0001_plain_morbius",
+      "breakpoints": true
     }
   ]
 }
